Extract utils into multiple files
